+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Changed
   * AZ-325: Upgrade keyvault module to latest version + Update CI
   * AZ-326: Add missing keyvault's inputs
 
+Updated
+  * AZ-410: Added a variable to enable choosing a TLS version for the logs storage account (Defaulted to TLS1_2)
+
 # v3.1.2/v4.0.0 - 2020-12-11
 
 Updated